Working on main. Main compiles but segfaults. Problem is in the Solve function. Just pushing to save my progress.

+++ SRC/quadSolver.h
@@ -7,7 +7,7 @@
 
 
 int mySquareRoot(double input, double *output); 
-int solve(float a, float b, float c, int * nroots, float *result[2]); // The Result of this 
+int solve(double a, double b, double c, int * nroots, double *result[2]); // The Result of this 
 void initLogfile(void);
 void logMessage(char *message);
 int validate(char *valuesString, double *result);
